New task search page controller+view+root component
