+++ SingularityUI/app/controllers/TaskSearchOld.coffee
