+++ SingularityUI/app/views/taskSearch.jsx
@@ -1,48 +1,20 @@
+import ReactView from './reactView';
+import TaskSearch from '../components/taskSearch/TaskSearch';
+
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import ReactView from './reactView';
+import { Provider } from 'react-redux';
 
-import TaskSearch from '../components/taskSearch/TaskSearch';
+export default class TaskSearchView extends ReactView {
 
-class TaskSearchView extends ReactView {
+  constructor(store, requestId) {
+    super();
+    this.store = store;
+    this.requestId = requestId;
+  }
 
-    constructor(...args) {
-        super(...args);
-        this.viewJson = this.viewJson.bind(this);
-    }
-
-    events() {
-        return _.extend(super.events(),
-            {'click [data-action="viewJSON"]': 'viewJson'});
-    }
-
-    viewJson(e) {
-        let $target = $(e.currentTarget).parents('tr');
-        let id = $target.data('id');
-        let collectionName = $target.data('collection');
-
-        // Need to reach into subviews to get the necessary data
-        let { collection } = this.subviews[collectionName];
-        return utils.viewJSON(collection.get(id));
-    }
-
-    initialize({requestId, global}, opts) {
-        this.requestId = requestId;
-        this.global = global;
-        this.opts = opts;
-    }
-
-    render() {
-      $(this.el).addClass("task-search-root");
-      ReactDOM.render(
-        <TaskSearch
-         initialRequestId = {this.requestId}
-         global = {this.global}
-         taskSearchViewSuper = {this.super}
-        />,
-      this.el);
+  render() {
+    ReactDOM.render(<Provider store={this.store}><TaskSearch requestId={this.requestId} /></Provider>, this.el);
   }
 }
-
-export default TaskSearchView;
